requirements.txt
# Counterfeit Medicine Detection — Deep Learning Dependencies
# Install with: pip install -r requirements.txt

tensorflow>=2.13.0
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
